controllers

+++ src/main/java/com/workintech/s18d4/controller/AccountController.java
@@ -1,40 +1,80 @@
 package com.workintech.s18d4.controller;
 
+import com.workintech.s18d4.dto.AccountResponse;
+import com.workintech.s18d4.dto.CustomerResponse;
 import com.workintech.s18d4.entity.Account;
+import com.workintech.s18d4.entity.Customer;
+import com.workintech.s18d4.service.AccountService;
+import com.workintech.s18d4.service.CustomerService;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@AllArgsConstructor
 @RestController
-@RequestMapping("/workintech/accounts")
+@RequestMapping("/account")
 public class AccountController {
-
-    @Autowired
     private AccountService accountService;
+    private CustomerService customerService;
 
     @GetMapping
-    public List<Account> getAllAccounts() {
+    public List<Account> findAll() {
         return accountService.findAll();
     }
 
-    @GetMapping("/{id}")
-    public Account getAccountById(@PathVariable Long id) {
-        return accountService.findById(id);
-    }
-
     @PostMapping("/{customerId}")
-    public Account createAccount(@PathVariable Long customerId, @RequestBody Account account) {
-        return accountService.save(customerId, account);
+    public AccountResponse save(@PathVariable("customerId") Long customerId, @RequestBody Account account) {
+
+        Customer customer = customerService.find(customerId);
+
+        if (customer != null) {
+            customer.getAccounts().add(account);
+            account.setCustomer(customer);
+            accountService.save(account);
+        } else {
+            throw new RuntimeException("Customer not found");
+        }
+
+        return new AccountResponse(account.getId(), account.getAccountName(), account.getMoneyAmount(),
+                new CustomerResponse(customer.getId(), customer.getEmail(), customer.getSalary()));
     }
 
-    @PutMapping("/{customerId}")
-    public Account updateAccount(@PathVariable Long customerId, @RequestBody Account account) {
-        return accountService.update(customerId, account);
+    @PutMapping
+    public AccountResponse update(@RequestBody Account account, @PathVariable("customerId") Long customerId) {
+        Customer customer = customerService.find(customerId);
+        Account toBeUpdatedAccount = null;
+
+
+        for (Account account1 : customer.getAccounts()) {
+            if (account.getId() == account1.getId()) {
+                toBeUpdatedAccount = account1;
+            }
+        }
+
+        if (toBeUpdatedAccount != null) {
+            throw new RuntimeException("Account(" + account.getId() + "not found");
+        }
+
+        int indexOfToBeUpdated = customer.getAccounts().indexOf(toBeUpdatedAccount);
+        customer.getAccounts().set(indexOfToBeUpdated, account);
+        account.setCustomer(customer);
+        accountService.save(account);
+        return new AccountResponse(account.getId(), account.getAccountName(), account.getMoneyAmount(),
+                new CustomerResponse(customer.getId(), customer.getEmail(), customer.getSalary()));
     }
 
     @DeleteMapping("/{id}")
-    public void deleteAccount(@PathVariable Long id) {
+    public AccountResponse remove(@PathVariable long id) {
+        Account account = accountService.find(id);
+       if(account == null){
+           throw new RuntimeException("Account not found");
+       }
+
         accountService.delete(id);
+        return new AccountResponse(account.getId(), account.getAccountName(), account.getMoneyAmount(),
+                new CustomerResponse(account.getCustomer().getId(), account.getCustomer().getEmail(), account.getCustomer().getSalary()));
     }
+
 }
